[API] Move `resolve_nuclio_version` to API
