+++ mlrun/api/crud/runtimes/nuclio/helpers.py
@@ -17,8 +17,11 @@ import urllib.parse
 import semver
 
 import mlrun
+import mlrun.api.utils.clients.nuclio
 import mlrun.api.utils.singletons.k8s
 import mlrun.runtimes
+from mlrun.config import config
+from mlrun.errors import err_to_str
 from mlrun.utils import logger
 
 
@@ -201,6 +204,29 @@ def is_nuclio_version_in_range(min_version: str, max_version: str) -> bool:
         )
         return True
     return parsed_min_version <= parsed_current_version < parsed_max_version
+
+
+# if nuclio version specified on mlrun config set it likewise,
+# if not specified, get it from nuclio api client
+# since this is a heavy operation (sending requests to API), and it's unlikely that the version
+# will change - cache it (this means if we upgrade nuclio, we need to restart mlrun to re-fetch the new version)
+def resolve_nuclio_version():
+    global cached_nuclio_version
+
+    if not cached_nuclio_version:
+
+        # config override everything
+        nuclio_version = config.nuclio_version
+        if not nuclio_version and config.nuclio_dashboard_url:
+            try:
+                nuclio_client = mlrun.api.utils.clients.nuclio.Client()
+                nuclio_version = nuclio_client.get_dashboard_version()
+            except Exception as exc:
+                logger.warning("Failed to resolve nuclio version", exc=err_to_str(exc))
+
+        cached_nuclio_version = nuclio_version
+
+    return cached_nuclio_version
 
 
 def compile_nuclio_archive_config(
